% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schema.R
\name{schema}
\alias{schema}
\title{Create a Random Design Schema}
\usage{
schema(Sites = NULL, NSubjects, BlockSize = NULL, RRatio = NULL, seed)
}
\arguments{
\item{Sites}{A number of sites / a list of site codes to be used}

\item{NSubjects}{The number of subjects assigned to each site}

\item{BlockSize}{The number of subjects assigned to each block}

\item{RRatio}{The randomization ratio used to assign treatment and control groups}

\item{seed}{The seed set for reproducibility}
}
\value{
A dataframe containing the design matrix
}
\description{
This function creates a completely randomized design of experiments
design schema in the form of a data.frame. It assumes that you know
what the various inputs of a designing a randomization schema are.
}
\section{Warning}{

Please don't use this function outside of the classroom
}

\author{
Matthias Quinn \href{mailto:miq_qedquinn@yahoo.com}{miq_qedquinn@yahoo.com}
}
